pyproject.toml
[tool.poetry]
name = "evremixes"
version = "4.1.0"
description = "Evanescence Remix Downloader"
authors = ["Danny Stewart <danny@stewart.cc>"]
readme = "README.md"
packages = [{include = "evremixes", from = "src"}]
include = ["*.json"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
dsutil = {git = "https://gitlab.dannystewart.com/danny/dsutil.git"}
halo = "^0.0.31"
inquirer = "^3.4.0"
mutagen = "^1.47.0"
pillow = "^11.1.0"
requests = "^2.32.3"
termcolor = "^2.5.0"
types-requests = "^2.32.0.20241016"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.1"
ruff = "^0.8.5"
mypy = "^1.14.1"

[tool.poetry.scripts]
evremixes = "evremixes.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
